backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
motor==3.3.2
pymongo==4.8.0
pydantic==2.6.0
python-multipart==0.0.6
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
dnspython==2.6.1


